rewrite: define PropTypes for core
